+++ mayan/apps/testing/tests/mixins.py
@@ -256,6 +256,11 @@ class RandomPrimaryKeyModelMonkeyPatchMixin:
 
         return primary_key
 
+    @classmethod
+    def setUpClass(cls):
+        super().setUpClass()
+        random.seed()
+
     def setUp(self):
         if self.random_primary_key_enable:
             self.method_save_original = models.Model.save